chore(debug): debug long running e2e
